Move launcherDir path
